specs: manually define debian_nextcodename when expected to be used
with puppet 4.x when you don't define the debian_nextcodename,
compilation fails saying that the variable is unknown. since we have
unit tests specifically for this fact's behaviour, we want to stub out
the fact's value in other tests so that we can verify the logic of
manifests is actually good.